Fixed: App fails to restart after changing settings bug
The bug was fixed by changing to the Mutex method of making the app
Single Instance only app instead of searching the running processes to
see if the app's process was running already.
The later method prevented the app from restarting because, when the app
attempts to restart the app is still in the running processes list and
thus prevents the app from restarting.